+++ src/main/java/com/sena/barberspa/repository/IProductoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sena.barberspa.model.Producto;
 
 @Repository
-public interface IProductoRepository extends JpaRepository<Producto, Integer> {
+public interface IProductoRepository extends JpaRepository<Producto, Long> {
 
 }
