+++ src/main/java/com/sena/barberspa/repository/IRecordatorioRepository.java
@@ -10,12 +10,12 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface IRecordatorioRepository extends JpaRepository<Recordatorio, Integer> {
+public interface IRecordatorioRepository extends JpaRepository<Recordatorio, Long> {
 
     List<Recordatorio> findByUsuarioAndActivoTrueOrderByFijadoDescFechaHoraAsc(Usuario usuario);
 
     @Query("SELECT r FROM Recordatorio r WHERE r.usuario = ?1 AND r.activo = true AND r.fechaHora BETWEEN ?2 AND ?3 ORDER BY r.fijado DESC, r.fechaHora ASC")
     List<Recordatorio> findProximosRecordatorios(Usuario usuario, LocalDateTime inicio, LocalDateTime fin);
 
-    List<Recordatorio> findByAgendamientoIdAndActivoTrue(Integer agendamientoId);
+    List<Recordatorio> findByAgendamientoIdAndActivoTrue(Long agendamientoId);
 }
