+++ src/main/java/com/sena/barberspa/service/IProductoService.java
@@ -9,11 +9,11 @@ public interface IProductoService {
 
 	public Producto save(Producto producto);
 
-	public Optional<Producto> get(Integer id);
+	public Optional<Producto> get(Long id);
 
 	public void update(Producto producto);
 
-	public void delete(Integer id);
+	public void delete(Long id);
 
 	public List<Producto> findAll();
